requirements.txt
flask
flask-socketio
eventlet
redis
opencv-python-headless
numpy
Pillow
ultralytics
